https://github.com/SeedCompany/url-redirect/issues/5
Note that the DNS entries need to be created by the script. They are not put in the redirects.yml file. The script needs to parse each redirect and determine what DNS entries are appropriate for that redirect. It will have to parse out the protocol, host, path, and queries itself. They won't be entered separately.

Since the port for HTTP and HTTPS is 80 and 443 respectively, we don't need to include that in the redirects.yml file.

This also means that the DNS entry needs to point to the ALB as a reference, not an IP address.

Assume all redirects are permanent. You don't need a field for the http response. Everything will be a 301.

Feel free to break this into smaller issues if you want. Thanks.